refactor: Move ParseMode to the parser module
